bug fixes + updates to many features

+++ backend/routes/decklistRoutes.js
@@ -12,7 +12,7 @@ decklistRouter.get('/', async (req, res) => {
   res.send(decklists);
 });
 
-const PAGE_SIZE = 10;
+const PAGE_SIZE = 20;
 
 decklistRouter.get(
   '/tournamentDeckSearch',
@@ -22,6 +22,7 @@ decklistRouter.get(
     const page = query.page || 1;
     const date = query.date || '';
     const tournament = query.tournament || '';
+    const order = query.order || '';
     const archetype = query.archetype || '';
 
     const dateFilter = date && date !== 'all' ? { date } : {};
@@ -30,6 +31,15 @@ decklistRouter.get(
     const archetypeFilter =
       archetype && archetype !== 'all' ? { archetype } : {};
 
+    const sortOrder =
+      order === 'featured'
+        ? { featured: -1 }
+        : order === 'newest'
+        ? { createdAt: -1 }
+        : order === 'oldest'
+        ? { createdAt: 1 }
+        : { _id: -1 };
+
     const decklists = await Decklist.find({
       ...archetypeFilter,
       ...tournamentFilter,
@@ -37,6 +47,7 @@ decklistRouter.get(
       isUserCreated: false,
     })
       .skip(pageSize * (page - 1))
+      .sort(sortOrder)
       .limit(pageSize);
 
     const countDecklists = await Decklist.countDocuments({
@@ -60,21 +71,20 @@ decklistRouter.get(
     const { query } = req;
     const id = query.id;
     const deck = await Decklist.findById(id);
-
     if (deck) {
       let prices = [];
       let mainDeckPrices = [];
       let extraDeckPrices = [];
       for (let index = 0; index < deck.mainDeck.length; index++) {
         const product = await Product.findOne({
-          slug: deck.mainDeck[index].slug,
+          name: deck.mainDeck[index].name.replace(/[^\x00-\x7F]/g, ''),
         });
         prices.push(product.price);
         mainDeckPrices.push(product.price);
       }
       for (let index = 0; index < deck.extraDeck.length; index++) {
         const product = await Product.findOne({
-          slug: deck.extraDeck[index].slug,
+          name: deck.extraDeck[index].name.replace(/[^\x00-\x7F]/g, ''),
         });
         prices.push(product.price);
         extraDeckPrices.push(product.price);
@@ -98,7 +108,6 @@ decklistRouter.get(
     const email = query.email || '';
     const deck = await Decklist.findById(id);
     const user = await User.findOne({ email });
-
     let yourPrice = 0.0;
 
     if (deck) {
@@ -107,42 +116,49 @@ decklistRouter.get(
       const counts = {};
       for (let index = 0; index < deck.mainDeck.length; index++) {
         const product = await Product.findOne({
-          slug: deck.mainDeck[index].slug,
+          name: deck.mainDeck[index].name.replace(/[^\x00-\x7F]/g, ''),
         });
+
         totalPrice = totalPrice + product.price;
-        counts[deck.mainDeck[index].slug] = counts[deck.mainDeck[index].slug]
-          ? counts[deck.mainDeck[index].slug] + 1
+        counts[deck.mainDeck[index].name] = counts[deck.mainDeck[index].name]
+          ? counts[deck.mainDeck[index].name] + 1
           : 1;
       }
       for (let index = 0; index < deck.extraDeck.length; index++) {
         const product = await Product.findOne({
-          slug: deck.extraDeck[index].slug,
+          name: deck.extraDeck[index].name.replace(/[^\x00-\x7F]/g, ''),
         });
         totalPrice = totalPrice + product.price;
-        counts[deck.extraDeck[index].slug] = counts[deck.extraDeck[index].slug]
-          ? counts[deck.extraDeck[index].slug] + 1
+        counts[deck.extraDeck[index].name] = counts[deck.extraDeck[index].name]
+          ? counts[deck.extraDeck[index].name] + 1
           : 1;
       }
-      totalPrice = totalPrice.toFixed(2);
 
+      totalPrice = totalPrice.toFixed(2);
       yourPrice = totalPrice;
       if (email !== '') {
         // loop through collection and subtract from totalPrice to get yourPrice
         for (const [key, value] of Object.entries(counts)) {
-          //console.log(`${key} : ${value}`);
-          let obj = user.cardCollection.find((o) => o.slug === key);
-          if (obj) {
+          let obj = user.cardCollection.find((o) => o.name === key);
+          if (obj !== undefined) {
             if (obj.count >= value) {
-              const product = await Product.findOne({ slug: obj.slug });
-              yourPrice = yourPrice - product.price * value;
+              const product = await Product.findOne({
+                name: obj.name.replace(/[^\x00-\x7F]/g, ''),
+              });
+              if (product.price) {
+                yourPrice = yourPrice - product.price * value;
+              }
             } else {
-              const product = await Product.findOne({ slug: obj.slug });
-              yourPrice = yourPrice - product.price * obj.count;
+              const product = await Product.findOne({
+                name: obj.name.replace(/[^\x00-\x7F]/g, ''),
+              });
+              if (product.price) {
+                yourPrice = yourPrice - product.price * obj.count;
+              }
             }
-          } else {
           }
         }
-        yourPrice = yourPrice.toFixed(2);
+        yourPrice = Math.round(yourPrice * 100) / 100;
       }
 
       res.send({
@@ -152,6 +168,83 @@ decklistRouter.get(
       });
     } else {
       res.status(404).send({ message: 'decklist not found' });
+    }
+  })
+);
+
+decklistRouter.get(
+  '/feature',
+  expressAsyncHandler(async (req, res) => {
+    const deck = await Decklist.findById('6420b82908dbda2fc8a49f67');
+    const secondDeck = await Decklist.findById('6420b82908dbda2fc8a49fd1');
+    let returnArray = [];
+
+    if (deck && secondDeck) {
+      returnArray.push(deck);
+      returnArray.push(secondDeck);
+      res.send({
+        decklists: returnArray,
+      });
+    } else {
+      res.send({
+        error: 'Decklists not found',
+      });
+    }
+  })
+);
+
+decklistRouter.get(
+  '/ownedCards',
+  expressAsyncHandler(async (req, res) => {
+    const { query } = req;
+    const id = query.id;
+    const email = query.email || '';
+    const deck = await Decklist.findById(id);
+    const user = await User.findOne({ email });
+
+    let tempCollection = user.cardCollection.slice();
+    let mainOwned = [];
+    let extraOwned = [];
+
+    if (deck) {
+      for (let index = 0; index < deck.mainDeck.length; index++) {
+        let findIndex = tempCollection.findIndex(
+          (o) => o.slug === deck.mainDeck[index].slug
+        );
+        if (findIndex !== -1) {
+          if (tempCollection[findIndex].count > 0) {
+            mainOwned.push(1);
+            tempCollection[findIndex].count =
+              tempCollection[findIndex].count - 1;
+          } else {
+            mainOwned.push(0);
+          }
+        } else {
+          mainOwned.push(0);
+        }
+      }
+      for (let index = 0; index < deck.extraDeck.length; index++) {
+        let findIndex = tempCollection.findIndex(
+          (o) => o.slug === deck.extraDeck[index].slug
+        );
+        if (findIndex !== -1) {
+          if (tempCollection[findIndex].count > 0) {
+            extraOwned.push(1);
+            tempCollection[findIndex].count =
+              tempCollection[findIndex].count - 1;
+          } else {
+            extraOwned.push(0);
+          }
+        } else {
+          extraOwned.push(0);
+        }
+      }
+      res.send({
+        mainOwned: mainOwned,
+        extraOwned: extraOwned,
+      });
+    } else {
+      res.sendStatus(404).message('Error in finding deck');
     }
   })
 );
@@ -211,9 +304,20 @@ decklistRouter.post(
   '/delete',
   expressAsyncHandler(async (req, res) => {
     const id = req.body.id;
+    const email = req.body.id;
 
-    await Decklist.deleteOne({ _id: id });
-    res.send({ message: 'Deck Deleted ' });
+    const user = await User.findOne({ email: email });
+    if (user) {
+      user.updateOne({
+        $pull: {
+          deckIDs: id,
+        },
+      });
+      await Decklist.deleteOne({ _id: id });
+      res.send({ message: 'Deck Deleted ' });
+    } else {
+      res.status(404).send({ message: 'User not found' });
+    }
   })
 );
 
@@ -226,38 +330,41 @@ decklistRouter.post(
     const isMainDeck = req.body.isMainDeck;
     const objID = new mongoose.Types.ObjectId();
 
-    if (isMainDeck) {
-      await Decklist.findOneAndUpdate(
-        {
-          _id: id,
-        },
-        {
-          $push: {
-            mainDeck: {
-              objID: objID,
-              name: name,
-              slug: slug,
-              count: 1,
-            },
+    if (objID) {
+      if (isMainDeck) {
+        await Decklist.findOneAndUpdate(
+          {
+            _id: id,
           },
-        }
-      );
-    } else {
-      await Decklist.findOneAndUpdate(
-        {
-          _id: id,
-        },
-        {
-          $push: {
-            extraDeck: {
-              objID: objID,
-              name: name,
-              slug: slug,
-              count: 1,
+          {
+            $push: {
+              mainDeck: {
+                objID: objID,
+                name: name,
+                slug: slug,
+                count: 1,
+              },
             },
+          }
+        );
+      } else {
+        await Decklist.findOneAndUpdate(
+          {
+            _id: id,
           },
-        }
-      );
+          {
+            $push: {
+              extraDeck: {
+                objID: objID,
+                name: name,
+                slug: slug,
+                count: 1,
+              },
+            },
+          }
+        );
+      }
+      res.send({ message: 'Update Success' });
     }
   })
 );
@@ -296,6 +403,7 @@ decklistRouter.post(
         }
       );
     }
+    res.send({ message: 'Card Deleted' });
   })
 );
 
@@ -319,6 +427,25 @@ decklistRouter.post(
 );
 
 decklistRouter.post(
+  '/changeArchetype',
+  expressAsyncHandler(async (req, res) => {
+    const archetype = req.body.archetype;
+    const id = req.body.id;
+
+    const decklist = await Decklist.findById(id);
+    if (decklist) {
+      decklist.archetype = archetype || decklist.archetype;
+      const updatedDecklist = await decklist.save();
+      res.send({
+        archetype: updatedDecklist.archetype,
+      });
+    } else {
+      res.status(404).send({ message: 'Error in changing archetype' });
+    }
+  })
+);
+
+decklistRouter.post(
   '/incrementCard',
   expressAsyncHandler(async (req, res) => {
     const slug = req.body.slug;
@@ -326,7 +453,6 @@ decklistRouter.post(
     const isMainDeck = req.body.isMainDeck;
 
     if (isMainDeck) {
-      console.log('updating main deck');
       await Decklist.findOneAndUpdate(
         {
           _id: id,
@@ -339,7 +465,6 @@ decklistRouter.post(
         }
       );
     } else {
-      console.log('updating extra deck');
       await Decklist.findOneAndUpdate(
         {
           _id: id,
